+++ NEWS.md
@@ -1,3 +1,7 @@
+# babelwhale 1.0.1
+
+* Fix issue with permanent saving of config.
+
 # babelwhale 1.0.0 (28-06-2019)
 
-* Release of babelwhale, a package to run docker and singularity containers from R
+* Release of babelwhale, a package to run docker and singularity containers from R.